requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.31.0
flask-limiter==3.5.0
python-dotenv==1.0.0 
werkzeug==2.0.1
google-generativeai==0.3.2